pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rss-podcast"
version = "0.1.0"
description = "RSS aggregator with AI-powered podcast generation"
authors = [
  { name = "Arkadiusz Miszczyszyn", email = "arkadiusz.miszczyszyn@gmail.com" },
]
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.10",
    "requests>=2.31.0",
    "trafilatura>=1.6.0",
    "lxml_html_clean>=0.1.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
tts = [
    "fish-speech>=0.1.0",
    "ormsgpack>=1.0.0",
    "requests>=2.31.0",
    "torch>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
rss-podcast = "rss_podcast.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]